+++ admin/statistics/database.py
@@ -1,7 +1,7 @@
 import logging
 from playhouse.shortcuts import model_to_dict
-from peewee import Model, SqliteDatabase, IntegerField, DateTimeField, FloatField, PrimaryKeyField, IntegrityError, \
-    DoesNotExist
+from peewee import (Model, SqliteDatabase, IntegerField, DateTimeField,
+                    FloatField, PrimaryKeyField, IntegrityError, DoesNotExist)
 from admin import DB_FILE_PATH
 
 logger = logging.getLogger(__name__)
